Refactor webpack configs
