+++ frontend/webpack/config.prod.js
@@ -7,20 +7,6 @@ const HtmlWebpackPlugin = require("html-webpack-plugin");
 
 const BASE_CONFIG = require("./config.base");
 
-const htmlWebpackPlugins = [
-  "terms-of-service",
-  "security-policy",
-  "privacy-policy",
-  "email-handler",
-  "das",
-].map((name) => {
-  return new HtmlWebpackPlugin({
-    filename: `${name}.html`,
-    template: resolve(__dirname, `../static/${name}.html`),
-    inject: false,
-  });
-});
-
 function pad(numbers, maxLength, fillString) {
   return numbers.map((number) =>
     number.toString().padStart(maxLength, fillString)
@@ -79,7 +65,6 @@ const PRODUCTION_CONFIG = {
       new CssMinimizerPlugin(),
     ],
   },
-  plugins: htmlWebpackPlugins,
 };
 
 module.exports = merge(BASE_CONFIG, PRODUCTION_CONFIG);
